Refactor RMMultiplicityValidation and RMOccurrenceValidation
