+++ tools/src/main/java/com/nedap/archie/rmobjectvalidator/RMObjectValidator.java
@@ -12,8 +12,6 @@ import com.nedap.archie.rminfo.InvariantMethod;
 import com.nedap.archie.rminfo.MetaModel;
 import com.nedap.archie.rminfo.ModelInfoLookup;
 import com.nedap.archie.rminfo.RMTypeInfo;
-import com.nedap.archie.rmobjectvalidator.validations.RMMultiplicityValidation;
-import com.nedap.archie.rmobjectvalidator.validations.RMOccurrenceValidation;
 import org.openehr.utils.message.I18n;
 
 import java.lang.reflect.InvocationTargetException;
@@ -36,8 +34,10 @@ public class RMObjectValidator extends RMObjectValidatingProcessor {
     private ReflectionConstraintImposer constraintImposer;
     private final ValidationConfiguration validationConfiguration;
     private boolean validateInvariants = true;
+    private final RmOccurrenceValidator rmOccurrenceValidator;
     private final RmPrimitiveObjectValidator rmPrimitiveObjectValidator;
     private final RmTupleValidator rmTupleValidator;
+    private final RmMultiplicityValidator rmMultiplicityValidator;
 
     /**
      * Creates an RM Object Validator with the given ModelInfoLook class, and the given OperationalTemplateProvider
@@ -59,8 +59,10 @@ public class RMObjectValidator extends RMObjectValidatingProcessor {
                 .failOnUnknownTerminologyId(com.nedap.archie.ValidationConfiguration.isFailOnUnknownTerminologyId())
                 .build();
         ValidationHelper validationHelper = new ValidationHelper(this.lookup, dummyValidationConfiguration);
+        rmOccurrenceValidator = new RmOccurrenceValidator();
         rmPrimitiveObjectValidator = new RmPrimitiveObjectValidator(validationHelper);
         rmTupleValidator = new RmTupleValidator(this.lookup, validationHelper, rmPrimitiveObjectValidator);
+        rmMultiplicityValidator = new RmMultiplicityValidator();
     }
 
     /**
@@ -77,8 +79,10 @@ public class RMObjectValidator extends RMObjectValidatingProcessor {
         this.validationConfiguration = validationConfiguration;
         ValidationHelper validationHelper = new ValidationHelper(this.lookup, validationConfiguration);
         validateInvariants = validationConfiguration.isValidateInvariants();
+        rmOccurrenceValidator = new RmOccurrenceValidator();
         rmPrimitiveObjectValidator = new RmPrimitiveObjectValidator(validationHelper);
         rmTupleValidator = new RmTupleValidator(this.lookup, validationHelper, rmPrimitiveObjectValidator);
+        rmMultiplicityValidator = new RmMultiplicityValidator();
     }
 
     /**
@@ -107,7 +111,7 @@ public class RMObjectValidator extends RMObjectValidatingProcessor {
     }
 
     private List<RMObjectValidationMessage> runArchetypeValidations(List<RMObjectWithPath> rmObjects, String path, CObject cobject) {
-        List<RMObjectValidationMessage> result = new ArrayList<>(RMOccurrenceValidation.validate(metaModel, rmObjects, path, cobject));
+        List<RMObjectValidationMessage> result = new ArrayList<>(rmOccurrenceValidator.validate(metaModel, rmObjects, path, cobject));
         if (rmObjects.isEmpty()) {
             //if this branch of the archetype tree is null in the reference model, we're done validating
             //this has to be done after validateOccurrences(), or required fields do not get validated
@@ -285,7 +289,7 @@ public class RMObjectValidator extends RMObjectValidatingProcessor {
 
         if (emptyObservationErrors.isEmpty()) {
 
-            result.addAll(RMMultiplicityValidation.validate(attribute, joinPaths(pathSoFar, "/", rmAttributeName), attributeValue));
+            result.addAll(rmMultiplicityValidator.validate(attribute, joinPaths(pathSoFar, "/", rmAttributeName), attributeValue));
 
             if(attribute.getChildren() == null || attribute.getChildren().isEmpty()) {
                 //no child CObjects. Cardinality/existence has already been validated. Run default RM validations
